requirements.txt
requests>=2.31.0
pandas>=2.1.0
pydantic-settings>=2.1.0
pyodbc>=5.1.0
azure-functions>=1.17.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0
python-dotenv>=1.0.0
